Rewrite UART and SPI drivers
